renamed commands classes
