+++ src/bot/message_command/HelpCommand.cpp
@@ -1,0 +1,9 @@
+#include "bot/command/HelpCommand.h"
+
+void HelpCommand::Execute(TgBot::Bot &bot, const TgBot::Message::Ptr &message) {
+    bot.getApi().sendMessage(message->chat->id,
+                             "Цей бот допомагає вам вести облік витрат і категорій, створювати фінансові звіти та"
+                             " аналізувати свої фінанси для кращого планування бюджету🤓📝\n"
+                             "Використовуйте цього бота для відстеження витрат, автоматичного підрахунку звітів та "
+                             "управління своїми фінансами, щоб досягти фінансової стабільності💪🏻✨");
+}
